change bill table item return

+++ src/main/java/com/outwork/accountingapiapp/models/entity/PosCardFeeEntity.java
@@ -27,5 +27,5 @@ public class PosCardFeeEntity {
     private PosEntity pos;
 
     @Column
-    private int posCardFee;
+    private double posCardFee;
 }
